Changed offset vector in data_member to be size_t instead of unsigned long. I'm not sure if size_t is ever not unsigned long, but at least this should take care of any of those cases

+++ src/data_member.hpp
@@ -17,7 +17,7 @@ class data_member {
 
   public:
 	std::string name; // The name of the data_member
-	std::vector<unsigned long>
+	std::vector<size_t>
 		offsets; // The memory offsets to get to the data member
 	std::type_info* type;
 	void str_to_type_id(std::string str);
